+++ src/Role.php
@@ -58,7 +58,7 @@ class Role extends Eloquent
      */
     public function users()
     {
-        return $this->belongsToMany(User::class, 'user_role')->withTimestamps();
+        return $this->belongsToMany(User::class, 'user_role', 'role_id', 'user_id')->withTimestamps();
     }
 
     /**